+++ app/src/test/java/com/readonchandler/view/readonchandler/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.lucky.readonchandler;
+package com.readonchandler.view.readonchandler;
 
 import org.junit.Test;
